// src/app/models/actualizacion.ts
import { environment } from "src/environments/environment";

export class Actualizacion {
    id: number;
    nombre: string;
    surname: string;
    especialidad: any;
    universidad: any;
    ano: string;
    org: string;
    website: string;
    email: string;
    direccion: string;
    direccion1: string;
    estado: string;
    ciudad: string;
    telefonos: string;
    tel1: string;
    telhome: string;
    telmovil: string;
    telprincipal: string;
    facebook: string;
    instagram: string;
    twitter: string;
    linkedin: string;
    vcard: string;
    created_at: Date;
    // status:string;


}

// const PUBLISHED = 'PUBLISHED';
//     const PENDING = 'PENDING';
//     const REJECTED = 'REJECTED';